requirements/prod-stor3.txt
-e git://github.com/GaretJax/coffin.git@9305129bec72776eddfed43ff3938e982144a08a#egg=Coffin-dev
-e git://github.com/django/django.git@4b722b31e1778310ba0c1ed96fdacf99aa203dd2#egg=Django-dev
Jinja2
MySQL-python
PyRRD
PyYAML
Pygments
-e hg+http://bitbucket.org/andrewgodwin/south/@c8fc82619ac9bb9549484a248cb16e65274da9d3#egg=South-dev
amqplib
anyjson
-e git://github.com/celery/celery.git@3077d7f0f1c227ab260e8a700fc5ca41bd2fada2#egg=celery-dev
celerymon
-e git://github.com/celery/django-celery.git@b14f639a96b5eb7dfed50eb1403d1065cc56544e#egg=django_celery-dev
django-assets
django-braces
django-compressor
django-crispy-forms
-e git://github.com/jbalogh/django-debug-cache-panel@2ff3c45be962e08f02928d021c61e7008964020a#egg=django_debug_cache_panel-dev
-e git://github.com/django-debug-toolbar/django-debug-toolbar.git@ae51b16d42a1ea52f99e1a89f1a74ec3c3cc6c20#egg=django_debug_toolbar-dev
-e git://github.com/playfire/django-debug-toolbar-user-panel.git@b212727fa38914fe9f41cca2a4ac5e9c0be93629#egg=django_debug_toolbar_user_panel-dev
django-extensions
django-indexer
-e git://github.com/alex2/django_logtail.git@9d12efe60a4151317de9f4e02200e298a51c5bd2#egg=django_logtail-dev
django-paging
django-supervisor
django-templatetag-sugar
ipaddr
ipcalc
ipdb
ipython
ipython-extensions
iterpipes
-e git://github.com/auzigog/jinja-bootstrap.git#egg=jinja-bootstrap
-e git://github.com/bradjasper/django-jsonfield.git@1306a07e6da629595bd03690e5bdf576a1e70f25#egg=jsonfield-dev
kombu
netbeacon
netifaces
python-flot-utils
-e git://github.com/joejulian/python-gluster.git@2f529f316f5833c334de051726f4bf6801b22628#egg=python_gluster-dev
pytz
raven
simplegeneric
simplejson
supervisor
threadpool
uWSGI
virtualenv
virtualenvwrapper

-e git://github.com/seb-m/pyinotify.git#egg=pyinotify
IPy

